+++ plugins/object.lua
@@ -5,6 +5,12 @@ Object = {}
 local abstraction_data = setmetatable({}, {__mode = "k"})
 
 local callbacks = {}
+local other_callbacks = {
+    "onCreate",
+    "onDestroy",
+    "onStep",
+    "onDraw"
+}
 
 
 
@@ -23,6 +29,23 @@ Object.ARRAY = {
 }
 
 
+Object.PARENT = {
+    actor               = gm.constants.pActor,
+    enemyClassic        = gm.constants.pEnemyClassic,
+    enemyFlying         = gm.constants.pEnemyFlying,
+    boss                = gm.constants.pBoss,
+    bossClassic         = gm.constants.pBossClassic,
+    pickupItem          = gm.constants.pPickupItem,
+    pickupEquipment     = gm.constants.pPickupEquipment,
+    drone               = gm.constants.pDrone,
+    mapObjects          = gm.constants.pMapObjects,
+    interactable        = gm.constants.pInteractable,
+    interactableChest   = gm.constants.pInteractableChest,
+    interactableCrate   = gm.constants.pInteractableCrate,
+    interactableDrone   = gm.constants.pInteractableDrone
+}
+
+
 Object.CUSTOM_START = 800
 
 
@@ -30,10 +53,8 @@ Object.CUSTOM_START = 800
 -- ========== Static Methods ==========
 
 Object.new = function(namespace, identifier, parent)
-    if Object.find(namespace, identifier) then
-        log.error("Object already exists", 2)
-        return nil
-    end
+    local obj = Object.find(namespace, identifier)
+    if obj then return obj end
 
     local obj = gm.object_add_w(namespace, identifier, Wrap.unwrap(parent))
     return Object.wrap(obj)
@@ -97,11 +118,7 @@ methods_object = {
     add_callback_obj_actual = function(self, callback, func)
         if self.value < Object.CUSTOM_START then return end
 
-        if callback == "onCreate"
-        or callback == "onDestroy"
-        or callback == "onStep"
-        or callback == "onDraw"
-        then
+        if Helper.table_has(other_callbacks, callback) then 
             local callback_id = self["on_"..string.lower(string.sub(callback, 3, 3))..string.sub(callback, 4, #callback)]
             if not callbacks[callback_id] then callbacks[callback_id] = {} end
             table.insert(callbacks[callback_id], func)
@@ -109,6 +126,14 @@ methods_object = {
         else log.error("Invalid callback name", 2)
 
         end
+    end,
+
+
+    clear_callbacks = function(self)
+        callbacks[self.on_create] = nil
+        callbacks[self.on_destroy] = nil
+        callbacks[self.on_step] = nil
+        callbacks[self.on_draw] = nil
     end,